Move Store interfaces to datastructures package
Remove core package dependency on storage package.
